+++ src/core/component/flyweight/index.ts
@@ -15,7 +15,7 @@ import { deprecate } from 'core/functools/deprecation';
 import Async from 'core/async';
 
 import { components } from 'core/component/const';
-import { supports, CreateElement, VNode } from 'core/component/engines';
+import { patchVNode, supports, CreateElement, VNode } from 'core/component/engines';
 
 import { initProps } from 'core/component/prop';
 import { initFields } from 'core/component/field';
@@ -26,7 +26,7 @@ import { implementEventAPI } from 'core/component/event';
 import { attachAccessorsFromMeta } from 'core/component/accessor';
 
 import { getNormalParent } from 'core/component/traverse';
-import { getComponentDataFromVNode, patchComponentVData } from 'core/component/vnode';
+import { getComponentDataFromVNode } from 'core/component/vnode';
 import { execRenderObject } from 'core/component/render';
 
 import { ComponentInterface } from 'core/component/interface';
@@ -202,8 +202,9 @@ export function parseVNodeAsFlyweight(
 	newVNode.fakeInstance = fakeCtx;
 	newVNode.data = newVNode.data ?? {};
 
-	patchComponentVData(newVNode.data, componentData, {
-		patchAttrs: Boolean(meta.params.inheritAttrs)
+	patchVNode(newVNode, fakeCtx, {
+		// @ts-ignore (unsafe cast)
+		data: componentData
 	});
 
 	// Attach component event listeners