+++ config/CHANGELOG.md
@@ -9,6 +9,22 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.137 (2021-02-04)
+
+#### :rocket: New Feature
+
+* Added a new parameter `--fat-html`
+
+#### :bug: Bug fix
+
+* Used `nanoid` to generate `nonceStore`
+
+## v3.0.0-rc.136 (2021-02-02)
+
+#### :rocket: New Feature
+
+* Added a new parameter `build.demoPage`
+
 ## v3.0.0-rc.115 (2020-12-23)
 
 #### :rocket: New Feature